+++ apps/omg_child_chain/lib/child_chain.ex
@@ -25,6 +25,7 @@ defmodule OMG.ChildChain do
   alias OMG.ChildChain.FeeServer
   alias OMG.ChildChain.FreshBlocks
   alias OMG.Fees
+  alias OMG.Fees.Filter
   alias OMG.State
   alias OMG.State.Transaction
   alias OMG.Status.Alert.Alarm
@@ -60,7 +61,7 @@ defmodule OMG.ChildChain do
           {:ok, Fees.full_fee_t()} | {:error, :currency_fee_not_supported}
   def get_filtered_fees(tx_types, currencies) do
     with {:ok, fees} <- FeeServer.transaction_fees() do
-      Fees.filter_fees(fees, tx_types, currencies)
+      Filter.filter(fees, tx_types, currencies)
     end
     |> result_with_logging()
   end
